ChickpeaAscoDispersal: Research Compendium for the Paper 'The role of conidia in the dispersal of Ascochyta rabiei'

Documents the workflow and processes involved in the analysis of data for the paper 'The role of conidia in the dispersal of Ascochyta rabiei'.

See also