ChickpeaAscoDispersal: Research Compendium for the Paper 'The role of conidia in the dispersal of Ascochyta rabiei'

Documents the workflow and processes involved in the analysis of data for the paper 'The role of conidia in the dispersal of Ascochyta rabiei', https://doi.org/10.1007/s10658-020-02126-2.

See also

Author

Maintainer: Adam H. Sparks adam.sparks@usq.edu.au (ORCID)

Authors:

Ihsanul Khaliq ihsan.khaliq@usq.edu.au (ORCID)

Joshua Fanning joshua.fanning@agriculture.vic.gov.au (ORCID)

Paul Melloy paul.melloy@usq.edu.au (ORCID)

Jean Galloway jean.galloway@dpird.wa.gov.au

Kevin Moore kevin.moore@dpi.nsw.gov.au (ORCID)

Daniel Burrell daniel.burrell@usq.edu.au (ORCID)

Other contributors:

GRDC Project USQ1903-003RTX [funder]

Agriculture Victoria and GRDC co-investment project DAV00150 [funder, data contributor]

GRDC Project DJP1097-001RTX [funder]

The Birchip Cropping Group [data contributor]